+++ javascript/UnixLepik.js
@@ -169,6 +169,10 @@ class UnixLepik extends LepikEvents_1.LepikEvents {
         const command = `xdotool windowminimize ${windowId}`;
         __classPrivateFieldGet(this, _UnixLepik_instances, "m", _UnixLepik_executeShellCommand).call(this, command);
     }
+    maximizeWindow(windowId) {
+        const command = `xdotool windowmaximize ${windowId}`;
+        __classPrivateFieldGet(this, _UnixLepik_instances, "m", _UnixLepik_executeShellCommand).call(this, command);
+    }
     /**
      * Delays the execution for the specified number of milliseconds.
      * @param {number} ms - The number of milliseconds to delay.